+++ src/useCases/calculateGrossSalary/CalculateGrossSalary.spec.ts
@@ -1,4 +1,4 @@
-import { AppError } from "../../erros/AppError";
+import { AppError } from "../../errors/AppError";
 import { CalculateGrossSalary } from "./CalculateGrossSalary";
 
 describe("Calculate gross salary", () => {
