+++ src/useCases/getWorkCoefficient/GetWorkCoefficient.spec.ts
@@ -1,4 +1,4 @@
-import { AppError } from "../../erros/AppError";
+import { AppError } from "../../errors/AppError";
 import { GetWorkCoefficient } from "./GetWorkCoefficient";
 
 describe("Get work coeficient", () => {
