+++ src/useCases/getWorkCoefficient/GetWorkCoefficient.ts
@@ -1,4 +1,4 @@
-import { AppError } from "../../erros/AppError";
+import { AppError } from "../../errors/AppError";
 
 type Coeficients = Record<string, number>;
